backend/requirements.txt
fastapi
uvicorn
Pillow
python-multipart
numpy
pydicom
torch
torchvision
opencv-python-headless 